+++ src/grepros/main.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     23.10.2021
-@modified    06.01.2022
+@modified    04.02.2022
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.main
@@ -409,7 +409,7 @@ def process_args(args):
     args.LINE_PREFIX = args.LINE_PREFIX and (args.RECURSE or len(args.FILES) != 1
                                              or args.PATHS or any("*" in x for x in args.FILES))
 
-    for k, v in vars(args).items():  # Flatten lists  of lists and drop duplicates
+    for k, v in vars(args).items():  # Flatten lists of lists and drop duplicates
         if k != "WRITE" and isinstance(v, list):
             here = set()
             setattr(args, k, [x for xx in v for x in (xx if isinstance(xx, list) else [xx])
@@ -485,6 +485,13 @@ def flush_stdout():
 
 def preload_plugins():
     """Imports and initializes plugins from auto-load folder and from arguments."""
+    plugins.add_write_format("bag", outputs.BagSink, "bag", [
+        ("overwrite=true|false",   "overwrite existing file in bag output\n"
+                                   "instead of appending to if bag or database\n"
+                                   "or appending unique counter to file name\n"
+                                   "(default false)")
+
+    ])
     args = make_parser().parse_known_args()[0] if "--plugin" in sys.argv else None
     try: plugins.init(process_args(args) if args else None)
     except ImportWarning: sys.exit(1)