+++ src/grepros/ros2.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     02.11.2021
-@modified    06.01.2022
+@modified    06.02.2022
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.ros2
@@ -27,6 +27,8 @@ import rclpy.duration
 import rclpy.executors
 import rclpy.serialization
 import rclpy.time
+import rosidl_parser.parser
+import rosidl_parser.definition
 import rosidl_runtime_py.utilities
 import yaml
 
@@ -49,11 +51,14 @@ ROS_TIME_CLASSES = {rclpy.time.Time:                 "builtin_interfaces/Time",
                     rclpy.duration.Duration:         "builtin_interfaces/Duration",
                     builtin_interfaces.msg.Duration: "builtin_interfaces/Duration"}
 
+## Mapping between type aliases and real types, like {"byte": "uint8"}
+ROS_ALIAS_TYPES = {"byte": "uint8", "char": "int8"}
+
 ## Data Distribution Service types to ROS builtins
 DDS_TYPES = {"boolean":             "bool",
              "float":               "float32",
              "double":              "float64",
-             "octet":               "int8",
+             "octet":               "byte",
              "short":               "int16",
              "unsigned short":      "uint16",
              "long":                "int32",
@@ -352,10 +357,19 @@ def canonical(typename):
     Converts DDS types like "octet" to "byte", and "sequence<uint8, 100>" to "uint8[100]".
     """
     is_array, bound, dimension = False, "", ""
-    match = re.match(r"sequence<([^\,>]+)\,?\s*(.*)>", typename)
+
+    match = re.match("sequence<(.+)>", typename)
     if match:  # "sequence<uint8, 100>" or "sequence<uint8>"
-        typename, is_array = match.group(1), True
-        if match.group(2): dimension = match.group(2)
+        is_array = True
+        typename = match.group(1)
+        match = re.match(r"([^,]+)?,\s?(\d+)", typename)
+        if match:  # sequence<uint8, 10>
+            typename = match.group(1)
+            if match.lastindex > 1: dimension = match.group(2)
+
+    match = re.match("(w?string)<(.+)>", typename)
+    if match:  # string<5>
+        typename, bound = match.groups()
 
     if "[" in typename:  # "string<=5[<=10]" or "string<=5[10]"
         dimension = typename[typename.index("[") + 1:typename.index("]")]
@@ -444,12 +458,15 @@ def get_message_type_hash(msg_or_type):
 
 @memoize
 def _get_message_definition(typename):
-    """Returns ROS2 message type definition full text (internal caching method)."""
+    """Returns ROS2 message type definition full text, or "" on error (internal caching method)."""
     try:
         texts, pkg = collections.OrderedDict(), typename.rsplit("/", 1)[0]
-        typepath = rosidl_runtime_py.get_interface_path(make_full_typename(typename))
-        with open(typepath) as f:
-            texts[typename] = f.read()
+        try:
+            typepath = rosidl_runtime_py.get_interface_path(make_full_typename(typename) + ".msg")
+            with open(typepath) as f:
+                texts[typename] = f.read()
+        except Exception:  # .msg file unavailable: parse IDL
+            texts[typename] = get_message_definition_idl(typename)
         for line in texts[typename].splitlines():
             if not line or not line[0].isalpha():
                 continue  # for line
@@ -460,12 +477,94 @@ def _get_message_definition(typename):
                        if "Header" == linetype else "%s/%s" % (pkg, linetype)
             linedef = None if linetype in texts else get_message_definition(linetype)
             if linedef: texts[linetype] = linedef
+
         basedef = texts.pop(next(iter(texts)))
         subdefs = ["%s\nMSG: %s\n%s" % ("=" * 80, k, v) for k, v in texts.items()]
-        return basedef + "\n".join(subdefs)
+        return basedef + ("\n" if subdefs else "") + "\n".join(subdefs)
     except Exception as e:
-        ConsolePrinter.error("Error reading type definition of %s: %s", typename, e)
+        ConsolePrinter.error("Error collecting type definition of %s: %s", typename, e)
         return ""
+
+
+@memoize
+def get_message_definition_idl(typename):
+    """
+    Returns ROS2 message type definition parsed from IDL file.
+
+    @since   version 0.4.2
+    """
+
+    def format_comment(text):
+        """Returns annotation text formatted with comment prefixes and escapes."""
+        ESCAPES = {"\n":   "\\n", "\t":   "\\t", "\x07": "\\a",
+                   "\x08": "\\b", "\x0b": "\\v", "\x0c": "\\f"}
+        repl = lambda m: ESCAPES[m.group(0)]
+        return "#" + "\n#".join(re.sub("|".join(map(re.escape, ESCAPES)), repl, l)
+                                for l in text.split("\\n"))
+
+    def format_type(typeobj, msgpackage, constant=False):
+        """Returns canonical type name, like "uint8" or "string<=5" or "nav_msgs/Path"."""
+        result = None
+        if isinstance(typeobj, rosidl_parser.definition.AbstractNestedType):
+            # Array, BoundedSequence, UnboundedSequence
+            valuetype = format_type(typeobj.value_type, msgpackage, constant)
+            size, bounding = "", ""
+            if isinstance(typeobj, rosidl_parser.definition.Array):
+                size = typeobj.size
+            elif typeobj.has_maximum_size():
+                size = typeobj.maximum_size
+            if isinstance(typeobj, rosidl_parser.definition.BoundedSequence):
+                bounding = "<="
+            result = "%s[%s%s]" % (valuetype, bounding, size) # type[], type[N], type[<=N]
+        elif isinstance(typeobj, rosidl_parser.definition.AbstractWString):
+            result = "wstring"
+        elif isinstance(typeobj, rosidl_parser.definition.AbstractString):
+            result = "string"
+        elif isinstance(typeobj, rosidl_parser.definition.NamespacedType):
+            nameparts = typeobj.namespaced_name()
+            result = canonical("/".join(nameparts))
+            if nameparts[0].value == msgpackage or "std_msgs/Header" == result:
+                result = canonical("/".join(nameparts[-1:]))  # Omit package if local or Header
+        else:  # Primitive like int8
+            result = DDS_TYPES.get(typeobj.typename, typeobj.typename)
+
+        if isinstance(typeobj, rosidl_parser.definition.AbstractGenericString) \
+        and typeobj.has_maximum_size() and not constant:  # Constants get parsed into "string<=N"
+            result += "<=%s" % typeobj.maximum_size
+
+        return result
+
+    def get_comments(obj):
+        """Returns all comments for annotatable object, as [text, ]."""
+        return [v.get("text", "") for v in obj.get_annotation_values("verbatim")
+                if "comment" == v.get("language")]
+
+    typepath = rosidl_runtime_py.get_interface_path(make_full_typename(typename) + ".idl")
+    with open(typepath) as f:
+        idlcontent = rosidl_parser.parser.parse_idl_string(f.read())
+    msgidl = idlcontent.get_elements_of_type(rosidl_parser.definition.Message)[0]
+    package = msgidl.structure.namespaced_type.namespaces[0]
+    DUMMY = rosidl_parser.definition.EMPTY_STRUCTURE_REQUIRED_MEMBER_NAME
+
+    lines = []
+    # Add general comments
+    lines.extend(map(format_comment, get_comments(msgidl.structure)))
+    # Add blank line between general comments and constants
+    if lines and msgidl.constants: lines.append("")
+    # Add constants
+    for c in msgidl.constants:
+        ctype = format_type(c.type, package, constant=True)
+        lines.extend(map(format_comment, get_comments(c)))
+        lines.append("%s %s=%s" % (ctype, c.name, c.value))
+    # Parser adds dummy placeholder if constants-only message
+    if not (len(msgidl.structure.members) == 1 and DUMMY == msgidl.structure[0].name):
+        # Add blank line between constants and fields
+        if msgidl.constants and msgidl.structure.members: lines.append("")
+        # Add fields
+        for m in msgidl.structure.members:
+            lines.extend(map(format_comment, get_comments(m)))
+            lines.append("%s %s" % (format_type(m.type, package), m.name))
+    return "\n".join(lines)
 
 
 @memoize
@@ -477,7 +576,7 @@ def _get_message_type_hash(typename):
 
 def get_message_fields(val):
     """Returns OrderedDict({field name: field type name}) if ROS2 message, else {}."""
-    if not is_ros_message(val): return val
+    if not is_ros_message(val): return {}
     fields = {k: canonical(v) for k, v in val.get_fields_and_field_types().items()}
     return collections.OrderedDict(fields)
 
@@ -489,10 +588,15 @@ def get_message_type(msg):
 
 def get_message_value(msg, name, typename):
     """Returns object attribute value, with numeric arrays converted to lists."""
-    v = getattr(msg, name)
+    v, scalartype = getattr(msg, name), scalar(typename)
     if isinstance(v, (bytes, array.array)) \
     or "numpy.ndarray" == "%s.%s" % (v.__class__.__module__, v.__class__.__name__):
-        return list(v)
+        v = list(map(int, v))
+    if v and isinstance(v, (list, tuple)) and scalartype in ("byte", "uint8"):
+        if isinstance(v[0], bytes):
+            v = list(map(ord, v))  # In ROS2, a byte array like [0, 1] is [b"\0", b"\1"]
+        elif scalartype == typename:
+            v = v[0]  # In ROS2, single byte values are given as bytes()
     return v
 
 
@@ -575,7 +679,7 @@ def qos_to_dict(qos):
 @memoize
 def scalar(typename):
     """
-    Returns scalar type from ROS2 message data type
+    Returns unbounded scalar type from ROS2 message data type
 
     Like "uint8" from "uint8[]", or "string" from "string<=10[<=5]".
     Returns type unchanged if not a collection or bounded type.