+++ CHANGELOG.rst
@@ -2,6 +2,20 @@
 Changelog for package grepros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.4.2 (2022-02-09)
+-------------------
+* add overwrite=true|false to --write options
+* add column-name=rostype:value to --write Parquet options
+* add scripts/generate_msgs.py
+* create message type definition from .idl if .msg file not available in ROS2
+* improve DDS type parsing in ROS2
+* handle ROS2 char and byte int8/uint8 reversal vs ROS1
+* fix assembling message type full definition in ROS2
+* fix processing byte values in ROS2
+* fix processing bounded string types in ROS2
+* fix inserting chars in Postgres
+* fix inserting very large integers in SQLite
+
 0.4.1 (2022-01-08)
 -------------------
 * refactor database and SQL sinks onto a common base